ext-stk: give test utils their own module
